% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_decontam_scores}
\alias{plot_decontam_scores}
\title{Plot the distribution of decontam scores}
\usage{
plot_decontam_scores(result, bins = 50L)
}
\arguments{
\item{result}{A result from \code{\link[=run_threshold_sweep]{run_threshold_sweep()}}.}

\item{bins}{Number of histogram bins.}
}
\value{
A \code{ggplot} object with candidate thresholds marked.
}
\description{
Plot the distribution of decontam scores
}
