% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-plots.R
\name{plot_sample_read_retention}
\alias{plot_sample_read_retention}
\title{Plot sample-level read retention as boxplots}
\usage{
plot_sample_read_retention(result, show_points = TRUE)
}
\arguments{
\item{result}{A result from \code{\link[=run_threshold_sweep]{run_threshold_sweep()}}.}

\item{show_points}{Logical; overlay individual samples.}
}
\value{
A ggplot object.
}
\description{
Plot sample-level read retention as boxplots
}
