% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-helpers.R
\name{summarize_library_sizes_by_group}
\alias{summarize_library_sizes_by_group}
\title{Summarize library sizes by sample group}
\usage{
summarize_library_sizes_by_group(ps, control_column, control_label)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{control_column}{Sample-data column identifying negative controls.}

\item{control_label}{Value or values identifying negative controls.}
}
\value{
A data frame containing group size and the standard library-size
summary statistics.
}
\description{
Summarize library sizes by sample group
}
